Cleaned some code organization
