+++ src/components/etablissement/etablissementRNCS/EtablissementRNCSGestion.vue
@@ -18,7 +18,7 @@ import EtablissementRNCSGestionMorale from '@/components/etablissement/etablisse
 import EtablissementRNCSGestionPhysique from '@/components/etablissement/etablissementRNCS/etablissementRNCSGestion/EtablissementRNCSGestionPhysique'
 import PanelNoResults from '@/components/etablissement/etablissementRNCS/templates/PanelNoResults'
 import Filters from '@/components/mixins/filters'
-import Formating from '@/components/etablissement/etablissementRNCS/mixins/RNCSformating'
+import RNCSFormating from '@/components/etablissement/etablissementRNCS/mixins/RNCSformating'
 
 export default {
   name: 'EtablissementRNCSGestion',
@@ -42,6 +42,6 @@ export default {
       ))
     }
   },
-  mixins: [Filters]
+  mixins: [Filters, RNCSFormating]
 }
 </script>
