+++ src/components/etablissement/etablissementRNCS/etablissementRNCSGestion/EtablissementRNCSGestionMorale.vue
@@ -23,7 +23,7 @@
 <script>
 import PanelInfo from '@/components/etablissement/etablissementRNCS/templates/PanelInfo'
 import Filters from '@/components/mixins/filters'
-import Formating from '@/components/etablissement/etablissementRNCS/mixins/RNCSformating'
+import RNCSFormating from '@/components/etablissement/etablissementRNCS/mixins/RNCSformating'
 
 export default {
   name: 'EtablissementRNCSGestionMorale',
@@ -47,6 +47,6 @@ export default {
       }
     }
   },
-  mixins: [Filters, Formating]
+  mixins: [Filters, RNCSFormating]
 }
 </script>
