+++ src/components/etablissement/etablissementSirene/EtablissementSireneContact.vue
@@ -14,6 +14,7 @@
 
 <script>
 import Filters from '@/components/mixins/filters'
+import Formating from '@/components/mixins/formating'
 
 export default {
   name: 'EtablissementSireneContact',
@@ -44,7 +45,7 @@ export default {
       return `${day}/${month}/${year}`
     }
   },
-  mixins: [Filters]
+  mixins: [Filters, Formating]
 }
 </script>
